requirements.txt
streamlit==1.28.0
spacy==3.7.2
pdfminer.six==20221105
pandas
